feature/IG-3-1 Реализована и доработана вёрстка Messages page
