+++ src/features/Messages/component/ChatBar/Chat/Chat.module.scss
@@ -1,0 +1,14 @@
+.chat {
+  height: 335px;
+  padding: 10px 20px 0 20px;
+  overflow: auto;
+}
+
+.itemMessage {
+  max-width: 220px;
+  border-radius: 50px;
+  background-color: #8e8e8e;
+  padding: 16px;
+  font-size: 14px;
+  word-break: break-word;
+}
