+++ src/features/Messages/component/ChatBar/ChatBar.module.scss
@@ -1,0 +1,14 @@
+.chatBar {
+  width: 585px;
+}
+
+.avatarHeaderChat {
+  width: 40px;
+  height: 40px;
+}
+
+.headerChat {
+  border-bottom: 1px solid #dbdbdb;
+  height: 60px;
+  padding: 0 20px;
+}
