+++ src/features/Messages/component/ChatBar/InputBar/InputBar.module.scss
@@ -1,0 +1,32 @@
+.inputBar {
+  padding: 20px;
+  border-radius: 100px;
+}
+
+.inputText {
+  outline: none;
+  border: none;
+  width: 100%;
+}
+
+.inputWrapper {
+  height: 35px;
+  border: 1px solid #dbdbdb;
+  border-radius: 40px;
+  padding: 0 10px;
+}
+
+button {
+  background-color: white;
+  border: none;
+}
+
+button:hover {
+  cursor: pointer;
+}
+
+.emojiPicker {
+  position: absolute;
+  z-index: 1;
+  margin: 0 0 365px -20px;
+}
