+++ src/features/Messages/component/ChatBar/InputBar/index.ts
@@ -1,0 +1,1 @@
+export { InputBar } from "./InputBar";
