+++ src/features/Messages/component/ChatBar/index.ts
@@ -1,0 +1,1 @@
+export { ChatBar } from "./ChatBar";
