+++ src/features/Messages/component/Messages.module.scss
@@ -4,112 +4,3 @@
   border: 1px solid #dbdbdb;
   background-color: white;
 }
-
-.usersBar {
-  width: 349px;
-  height: 483px;
-  border-right: 1px solid #dbdbdb;
-}
-
-.headerUserBar {
-  border-bottom: 1px solid #dbdbdb;
-  height: 60px;
-  padding: 0 20px;
-}
-
-.barMessageItem {
-  margin-top: 10px;
-}
-
-.userMessageItem {
-  height: 55px;
-  padding: 8px 20px;
-}
-
-.userMessageItem:active {
-  background-color: #dbdbdb;
-}
-
-.active {
-  background-color: #dbdbdb;
-}
-
-.userMessageItem:hover {
-  cursor: pointer;
-}
-
-.avatar {
-  width: 56px;
-  height: 56px;
-}
-
-.avatarHeaderChat{
-  width: 40px;
-  height: 40px;
-}
-
-.lastMessageText {
-  opacity: 0.5;
-}
-
-.headerChat {
-  border-bottom: 1px solid #dbdbdb;
-  height: 60px;
-  padding: 0 20px;
-}
-
-.chatBar {
-  width: 585px;
-}
-
-.invBlock {
-  width: 15px;
-}
-
-.chat {
-  height: 335px;
-  padding: 10px 20px 0 20px;
-  overflow: auto;
-}
-
-.inputBar {
-  padding: 20px;
-  border-radius: 100px;
-}
-
-.inputText {
-  outline: none;
-  border: none;
-  width: 100%;
-}
-
-.inputWrapper {
-  height: 35px;
-  border: 1px solid #dbdbdb;
-  border-radius: 40px;
-  padding: 0 10px;
-}
-
-button{
-  background-color: white;
-  border: none;
-}
-
-button:hover{
-  cursor: pointer;
-}
-
-.emojiPicker{
-  position: absolute;
-  z-index: 1;
-  margin: 0 0 365px -20px;
-}
-
-.itemMessage{
-  max-width: 220px;
-  border-radius: 50px;
-  background-color: #8e8e8e;
-  padding: 16px;
-  font-size: 14px;
-  word-break: break-word;
-}
