+++ src/features/Messages/component/UserBarMessages/index.ts
@@ -1,0 +1,1 @@
+export { UserBarMessages } from "./UserBarMessages";
